refactor: move betAmountInupt to component #184
