+++ packages/ui-react/src/components/game/BettingPanel.tsx
@@ -1,22 +1,15 @@
-import {
-  CASINO_GAME_TYPE,
-  FORMAT_TYPE,
-  formatRawAmount,
-  GAS_TOKEN_ADDRESS,
-} from "@betswirl/sdk-core"
+import { CASINO_GAME_TYPE, formatRawAmount, GAS_TOKEN_ADDRESS } from "@betswirl/sdk-core"
 import { WalletModal } from "@coinbase/onchainkit/wallet"
-import Decimal from "decimal.js"
-import { ChangeEvent, useEffect, useRef, useState } from "react"
-import { parseUnits } from "viem"
+import { useEffect, useRef, useState } from "react"
 import { useChain } from "../../context/chainContext"
 import { useBetRequirements } from "../../hooks/useBetRequirements"
 import { cn } from "../../lib/utils"
 import { BetStatus, ChainTokenPanelView, TokenWithImage } from "../../types/types"
 import { Button } from "../ui/button"
 import { ChainIcon } from "../ui/ChainIcon"
-import { Label } from "../ui/label"
 import { Sheet } from "../ui/sheet"
 import { TokenIcon } from "../ui/TokenIcon"
+import { BetAmountInput } from "./BetAmountInput"
 import { ChainAndTokenSheetPanel } from "./ChainAndTokenSheetPanel"
 
 interface BettingPanelProps {
@@ -44,8 +37,6 @@ interface BettingPanelProps {
   isMounted: boolean
 }
 
-const BET_AMOUNT_INPUT_STEP = 0.0001
-
 export function BettingPanel({
   game,
   balance,
@@ -71,21 +62,9 @@ export function BettingPanel({
   isMounted,
 }: BettingPanelProps) {
   const { appChainId, switchWalletChain } = useChain()
-  const [inputValue, setInputValue] = useState<string>("")
-  const [isValidInput, setIsValidInput] = useState<boolean>(true)
-  const [isUserTyping, setIsUserTyping] = useState<boolean>(false)
   const [isChainTokenSheetOpen, setIsChainTokenSheetOpen] = useState<boolean>(false)
   const [panelInitialView, setPanelInitialView] = useState<ChainTokenPanelView>("main")
-  const typingTimeoutRef = useRef<NodeJS.Timeout | null>(null)
   const [isWalletModalOpen, setIsWalletModalOpen] = useState(false)
-
-  useEffect(() => {
-    return () => {
-      if (typingTimeoutRef.current) {
-        clearTimeout(typingTimeoutRef.current)
-      }
-    }
-  }, [])
 
   // Track previous values to detect actual changes
   const prevChainIdRef = useRef(appChainId)
@@ -98,27 +77,11 @@ export function BettingPanel({
 
     if (chainChanged || tokenChanged) {
       onBetAmountChange(undefined)
-      setInputValue("")
-      setIsValidInput(true)
     }
 
     prevChainIdRef.current = appChainId
     prevTokenAddressRef.current = token.address
   }, [appChainId, token.address, onBetAmountChange]) // eslint-disable-line react-hooks/exhaustive-deps
-
-  // Sync input value with betAmount when user is not typing
-  useEffect(() => {
-    if (isUserTyping) return
-
-    if (betAmount === undefined) {
-      setInputValue("")
-      setIsValidInput(true)
-    } else {
-      const formatted = formatRawAmount(betAmount, token.decimals, FORMAT_TYPE.PRECISE)
-      setInputValue(formatted)
-      setIsValidInput(true)
-    }
-  }, [betAmount, token.decimals, isUserTyping])
 
   const {
     isAllowed: isTokenAllowed,
@@ -267,41 +230,6 @@ export function BettingPanel({
     }
   }
 
-  const handleInputChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const newInputValue = e.target.value
-
-    setInputValue(newInputValue)
-    setIsUserTyping(true)
-
-    if (typingTimeoutRef.current) {
-      clearTimeout(typingTimeoutRef.current)
-    }
-
-    typingTimeoutRef.current = setTimeout(() => {
-      setIsUserTyping(false)
-    }, 1000)
-
-    if (newInputValue === "") {
-      onBetAmountChange(undefined)
-      setIsValidInput(true)
-      return
-    }
-
-    try {
-      new Decimal(newInputValue)
-
-      try {
-        const weiValue = parseUnits(newInputValue, token.decimals)
-        onBetAmountChange(weiValue)
-        setIsValidInput(true)
-      } catch (_error) {
-        setIsValidInput(false)
-      }
-    } catch (_error) {
-      setIsValidInput(false)
-    }
-  }
-
   return (
     <div className="bg-control-panel-background p-4 rounded-[16px] flex flex-col gap-4">
       <div className="flex flex-col gap-3">
@@ -325,48 +253,14 @@ export function BettingPanel({
           </Button>
         </div>
 
-        <Label
-          htmlFor="betAmount"
-          className="text-sm font-medium -mb-1 text-text-on-surface-variant"
-        >
-          Bet amount
-        </Label>
-        <div className="relative flex h-12 w-full items-center text-sm">
-          <input
-            id="betAmount"
-            type="number"
-            placeholder="0"
-            min={0}
-            max={Number.parseFloat(formattedBalance)}
-            step={BET_AMOUNT_INPUT_STEP}
-            value={inputValue}
-            onChange={handleInputChange}
-            disabled={isInputDisabled}
-            className={cn(
-              "flex h-full w-full rounded-[12px] border-0",
-              "bg-neutral-background text-foreground font-semibold",
-              "px-4 py-2 pr-16",
-              "text-base placeholder:text-muted-foreground",
-              "ring-offset-background focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-primary",
-              "disabled:cursor-not-allowed disabled:opacity-50",
-              !isValidInput && "text-muted-foreground",
-            )}
-          />
-          <Button
-            variant="ghost"
-            onClick={handleTokenClick}
-            className={cn(
-              "absolute right-[12px] top-1/2 -translate-y-1/2 transform",
-              "flex items-center text-foreground font-medium gap-1",
-              "h-auto w-fit p-0 bg-transparent hover:bg-transparent hover:opacity-80 transition-opacity",
-              "border-0 shadow-none outline-none focus:outline-none",
-            )}
-            disabled={isInputDisabled}
-          >
-            <TokenIcon token={token} size={18} />
-            <span>{token.symbol}</span>
-          </Button>
-        </div>
+        <BetAmountInput
+          betAmount={betAmount}
+          onBetAmountChange={onBetAmountChange}
+          token={token}
+          isDisabled={isInputDisabled}
+          onTokenClick={handleTokenClick}
+          formattedBalance={formattedBalance}
+        />
 
         <div className="grid grid-cols-3 gap-2">
           <Button
